Migrate screens/models menu
