+++ lib/features/menu/menu_model.dart
@@ -1,0 +1,19 @@
+class MenuItem {
+  final String title;
+  final String description;
+  final double price;
+  final String image;
+  final String category;
+
+
+  /* Déclaration d'un constructeur */
+  MenuItem({
+    required this.title,
+    required this.description,
+    required this.price,
+    required this.image,
+    required this.category,
+  });
+
+
+}
