--- lib/model/menu_item.dart
@@ -1,19 +1,0 @@
-class MenuItem {
-  final String title;
-  final String description;
-  final double price;
-  final String image;
-  final String category;
-
-
-  /* Déclaration d'un constructeur */
-  MenuItem({
-    required this.title,
-    required this.description,
-    required this.price,
-    required this.image,
-    required this.category,
-  });
-
-
-}
