--- lib/screen/menu_screen/menu_detail.dart
@@ -1,1 +1,0 @@
-
